requirements.txt
requests
secret
bs4
python-dotenv